Create Positioning Attempt #1
This is our first real attempt at a positioning program based on encoders and a gyro. It uses the gyro as a rate gyro, and ignores anything under a certain value, attributing that to drift. It also uses the average of the left and right encoders, since that gives the position of the center of the robot.